refactor(addstatus.ts): addStatus を外だし
