+++ src/addStatus.ts
@@ -1,0 +1,35 @@
+import {
+  Graph,
+  ChangeStatus,
+} from '@ysk8hori/typescript-graph/dist/src/models';
+
+export default function addStatus(
+  {
+    modified,
+    created,
+    deleted,
+  }: {
+    modified: string[];
+    created: string[];
+    deleted: string[];
+  },
+  graph: Graph,
+): Graph {
+  const { nodes, relations } = graph;
+  const newNodes = nodes.map(node => {
+    const changeStatus: ChangeStatus = (function () {
+      if (deleted.includes(node.path)) return 'deleted';
+      if (created.includes(node.path)) return 'created';
+      if (modified.includes(node.path)) return 'modified';
+      return 'not_modified';
+    })();
+    return {
+      ...node,
+      changeStatus,
+    };
+  });
+  return {
+    nodes: newNodes,
+    relations,
+  };
+}
